Drop shorthand parser hook
It will be easier to implement proper logic in
an custom PreparedResponse, as there may be
cases where we will need to access additional
information about the request. In that case,
the :meth:`PreparedRequest.prepare` method
has access to all relevant data of an api
call. It should therefore be extended,
instead of this hook implemented.